writeups/dockerlabs/scripts/gallery/requirements.txt
requests
beautifulsoup4
pwntools
sshtunnel
